Moved provider to separate module
